backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
MarkupSafe==3.0.3
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
threadpoolctl==3.6.0
tifffile==2025.5.10
Werkzeug==3.1.6
